Update best_answer_id on question when answe is removed

+++ database/migrations/2021_02_03_100000_create_questions_table.php
@@ -14,6 +14,7 @@ class CreateQuestionsTable extends Migration
     public function up()
     {
         Schema::create('questions', function (Blueprint $table) {
+          //  $table->engine = "InnoDB";
             $table->id();
             $table->string('title');
             $table->string('slug')->unique();
@@ -21,9 +22,10 @@ class CreateQuestionsTable extends Migration
             $table->unsignedInteger('views')->default(0);
             $table->unsignedInteger('answers_count')->default(0);
             $table->integer('votes')->default(0);
-            $table->unsignedInteger('best_answer_id')->nullable();
+            $table->unsignedBigInteger('best_answer_id')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
+           // $table->foreign('best_answer_id')->references('id')->on('answers')->onDelete('SET NULL');
             $table->timestamps();
         });
     }
